moved stubs and hooks to mir_to_goto
