+++ compiler/rustc_codegen_llvm/src/gotoc/mir_to_goto/mod.rs
@@ -1,8 +1,9 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 
-pub mod monomorphize;
-//pub mod stubs;
 pub mod metadata;
+pub mod monomorphize;
 pub use metadata::GotocCtx;
 pub mod codegen;
+pub mod hooks;
+pub mod stubs;
